Migrate from `Constrained10` to `Entails`.
`Constrained10` still exists, but now in a simpler form and only as a
means to provide `Entails` for `Field10`.  Functions that provide an
instance now do so by way of `Entails` plus `FunctorWithIndex` et al.

Now hand-written GADTs that know how to provide instances can
participate in the fun.  Sadly, they can't participate in the Generics.

Change-Id: I956cbfbfb837f49f601e102f12f0221d0522037d